fix: refactor unstruct DL profile naming
